Update diagrams and code analysis reports
