+++ .gitignore
@@ -3,3 +3,4 @@ backend/.env
 AGENTS.md
 docker/build_commands.txt
 frontend/frontend.iml
+/.scannerwork
