+++ backend/src/main/java/com/tfg/backend/controller/ShopRestController.java
@@ -35,7 +35,7 @@ public class ShopRestController {
     private final ShopUserOrchestrator shopUserOrchestrator;
 
     // Inject the user connection service for presence enrichment
-    private final UserConnectionService userConnectionService;
+    private final ConnectionService connectionService;
 
     @Operation(summary = "(Manager) Get lightweight list of managed shops (Key-Value)")
     @GetMapping("/references")
@@ -181,7 +181,7 @@ public class ShopRestController {
     private ShopDTO toEnrichedDTO(Shop shop) {
         ShopDTO dto = new ShopDTO(shop);
         if (dto.getAssignedManager() != null) {
-            userConnectionService.enrichWithConnection(dto.getAssignedManager());
+            connectionService.enrichWithConnection(dto.getAssignedManager());
         }
         return dto;
     }
@@ -197,7 +197,7 @@ public class ShopRestController {
                 .filter(Objects::nonNull)
                 .toList();
 
-        userConnectionService.enrichWithConnections(managers);
+        connectionService.enrichWithConnections(managers);
         return dtos;
     }
 
@@ -214,7 +214,7 @@ public class ShopRestController {
                 .filter(Objects::nonNull)
                 .toList();
 
-        userConnectionService.enrichWithConnections(managers);
+        connectionService.enrichWithConnections(managers);
 
         // 3. Return the formatted PageResponse
         return PageFormatter.toPageResponse(dtoPage, dto -> dto);
@@ -227,7 +227,7 @@ public class ShopRestController {
     private TruckDTO toEnrichedTruckDTO(Truck truck) {
         TruckDTO dto = new TruckDTO(truck);
         if (dto.getAssignedDriver() != null) {
-            userConnectionService.enrichWithConnection(dto.getAssignedDriver());
+            connectionService.enrichWithConnection(dto.getAssignedDriver());
         }
         return dto;
     }
