+++ backend/src/main/java/com/tfg/backend/repository/ConnectionRepository.java
@@ -1,0 +1,9 @@
+package com.tfg.backend.repository;
+
+import com.tfg.backend.model.Connection;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ConnectionRepository extends MongoRepository<Connection, String> {
+}
