+++ docs/pages/01-features.md
@@ -2,6 +2,15 @@
 
 ## 🔗 Common Features and Screens
 
+**Dynamic Breadcrumbs and Reload Bar**
+
+Check your current location in the system at any moment, go back to previous pages or quickly refresh current page data.
+
+![Dynamic Breadcrumbs and Reload Bar](/docs/images/v0.2/breadcrumb-reload.png)
+
+&nbsp;
+
+
 **Login Page**
 
 Access the application using your credentials or your Google account, or start the password recovery process if you have forgotten it.
@@ -24,11 +33,24 @@ Regain access to your account by entering a one-time password (OTP) sent to your
 &nbsp;
 
 
-**Notifications Page**
+**Notifications Page and Notifications Tray**
 
 Get real-time updates about all different events depending on your role: order updates, assignment change, new products offering, etc.
 
 ![Notifications Page](/docs/images/v0.2/notifications.png)
+
+![Notifications Tray](/docs/images/v0.2/notifications-tray.png)
+
+&nbsp;
+
+
+**Connection Status Section and Online Indicators**
+
+Check current system status, last session insights (date, time and duration) and other users connection status (online / offline) across the app.
+
+![Connection Status Section](/docs/images/v0.2/connection-section.png)
+
+![Connection Status Indicators](/docs/images/v0.2/connection-indicator.png)
 
 &nbsp;
 
@@ -38,7 +60,7 @@ Get real-time updates about all different events depending on your role: order u
 - [All detailed views](/docs/pages/01-shop-subsystem.md)
 - Screen flow diagram:
 
-![Shop Subsystem Screens Diagram](/docs/images/v0.1/Frict-Client.png)
+![Shop Subsystem Screens Diagram](/docs/images/v0.1/Frict-0.2-Client.png)
 
 &nbsp;
 
@@ -47,7 +69,7 @@ Get real-time updates about all different events depending on your role: order u
 - [All detailed views](/docs/pages/01-management-subsystem.md)
 - Screen flow diagram:
 
-![Management Subsystem Screens Diagram](/docs/images/v0.1/Frict-Management.png)
+![Management Subsystem Screens Diagram](/docs/images/v0.1/Frict-0.2-Management.png)
 
 
 &nbsp;
